fix typo bugs, merge DotProto.Generate.Common into DotProto.Internal
