+++ proto3-suite.cabal
@@ -26,7 +26,6 @@ library
                        Proto3.Suite.Class
                        Proto3.Suite.DotProto
                        Proto3.Suite.DotProto.Generate
-                       Proto3.Suite.DotProto.Generate.Common
                        Proto3.Suite.DotProto.AST
                        Proto3.Suite.DotProto.Parsing
                        Proto3.Suite.DotProto.Rendering
